Major project cleanup and documentation overhaul
- Remove deprecated meta_cognition.rs and sequential.rs modules
- Add comprehensive reasoning engine with session management
- Implement new MCP tools: reasoning_suite, translator, streaming
- Add extensive test coverage for MCP protocol compliance
- Clean up old test files and consolidate test structure
- Add schema definitions for MCP tools and reasoning operations
- Implement proper stdio handshake protocol
- Add SQLite storage adapters for reasoning sessions
- Update build configuration and dependencies

+++ tests/vectorstore_vectorindex_tests.rs
@@ -58,7 +58,15 @@ fn test_vectorindex_search_cosine_similarity() -> Result<()> {
     store.add(2, vec![0.0, 1.0, 0.0, 0.0])?;
 
     // Vector 3: [0, 0.7071, 0.7071, 0] - similar to 2
-    store.add(3, vec![0.0, f64::consts::FRAC_1_SQRT_2, f64::consts::FRAC_1_SQRT_2, 0.0])?;
+    store.add(
+        3,
+        vec![
+            0.0,
+            std::f64::consts::FRAC_1_SQRT_2 as f32,
+            std::f64::consts::FRAC_1_SQRT_2 as f32,
+            0.0,
+        ],
+    )?;
 
     // Query with vector similar to 2 and 3
     let query = vec![0.0, 1.0, 0.0, 0.0];
